+++ packages/kit/src/utils/routing.spec.js
@@ -1,5 +1,5 @@
 import { assert, expect, test } from 'vitest';
-import { exec, parse_route_id, resolve_entry } from './routing.js';
+import { exec, parse_route_id } from './routing.js';
 
 const tests = {
 	'/': {
@@ -220,56 +220,4 @@ for (const { path, route, expected } of exec_tests) {
 test('parse_route_id errors on bad param name', () => {
 	assert.throws(() => parse_route_id('abc/[b-c]'), /Invalid param: b-c/);
 	assert.throws(() => parse_route_id('abc/[bc=d-e]'), /Invalid param: bc=d-e/);
-});
-
-const from_entry_tests = [
-	{
-		route: '/blog/[one]/[two]',
-		entry: { one: 'one', two: 'two' },
-		expected: '/blog/one/two'
-	},
-	{
-		route: '/blog/[one=matcher]/[...two]',
-		entry: { one: 'one', two: 'two/three' },
-		expected: '/blog/one/two/three'
-	},
-	{
-		route: '/blog/[one=matcher]/[[two]]',
-		entry: { one: 'one' },
-		expected: '/blog/one'
-	},
-	{
-		route: '/blog/[one]/[two]-and-[three]',
-		entry: { one: 'one', two: '2', three: '3' },
-		expected: '/blog/one/2-and-3'
-	},
-	{
-		route: '/blog/[one]/[...two]-not-three',
-		entry: { one: 'one', two: 'two/2' },
-		expected: '/blog/one/two/2-not-three'
-	}
-];
-
-for (const { route, entry, expected } of from_entry_tests) {
-	test(`resolve_entry generates correct path for ${route}`, () => {
-		const result = resolve_entry(route, entry);
-		assert.equal(result, expected);
-	});
-}
-
-test('resolve_entry errors on missing entry for required param', () => {
-	expect(() => resolve_entry('/blog/[one]/[two]', { one: 'one' })).toThrow(
-		"Missing parameter 'two' in route /blog/[one]/[two]"
-	);
-});
-
-test('resolve_entry errors on entry values starting or ending with slashes', () => {
-	assert.throws(
-		() => resolve_entry('/blog/[one]/[two]', { one: 'one', two: '/two' }),
-		"Parameter 'two' in route /blog/[one]/[two] cannot start or end with a slash -- this would cause an invalid route like foo//bar"
-	);
-	assert.throws(
-		() => resolve_entry('/blog/[one]/[two]', { one: 'one', two: 'two/' }),
-		"Parameter 'two' in route /blog/[one]/[two] cannot start or end with a slash -- this would cause an invalid route like foo//bar"
-	);
 });
